+++ services/vod/describe_vod_refresh_tasks.go
@@ -94,9 +94,9 @@ type DescribeVodRefreshTasksRequest struct {
 type DescribeVodRefreshTasksResponse struct {
 	*responses.BaseResponse
 	RequestId  string `json:"RequestId" xml:"RequestId"`
-	PageNumber int    `json:"PageNumber" xml:"PageNumber"`
-	PageSize   int    `json:"PageSize" xml:"PageSize"`
-	TotalCount int    `json:"TotalCount" xml:"TotalCount"`
+	PageNumber int64  `json:"PageNumber" xml:"PageNumber"`
+	PageSize   int64  `json:"PageSize" xml:"PageSize"`
+	TotalCount int64  `json:"TotalCount" xml:"TotalCount"`
 	Tasks      Tasks  `json:"Tasks" xml:"Tasks"`
 }
