+++ services/vod/struct_audio.go
@@ -22,13 +22,13 @@ type Audio struct {
 	Title             string     `json:"Title" xml:"Title"`
 	Tags              string     `json:"Tags" xml:"Tags"`
 	Status            string     `json:"Status" xml:"Status"`
-	Size              int        `json:"Size" xml:"Size"`
+	Size              int64      `json:"Size" xml:"Size"`
 	Duration          float64    `json:"Duration" xml:"Duration"`
 	Description       string     `json:"Description" xml:"Description"`
 	ModificationTime  string     `json:"ModificationTime" xml:"ModificationTime"`
 	CreationTime      string     `json:"CreationTime" xml:"CreationTime"`
 	CoverURL          string     `json:"CoverURL" xml:"CoverURL"`
-	CateId            int        `json:"CateId" xml:"CateId"`
+	CateId            int64      `json:"CateId" xml:"CateId"`
 	CateName          string     `json:"CateName" xml:"CateName"`
 	DownloadSwitch    string     `json:"DownloadSwitch" xml:"DownloadSwitch"`
 	PreprocessStatus  string     `json:"PreprocessStatus" xml:"PreprocessStatus"`
