+++ services/vod/struct_material.go
@@ -21,7 +21,7 @@ type Material struct {
 	Title        string                                `json:"Title" xml:"Title"`
 	Tags         string                                `json:"Tags" xml:"Tags"`
 	Status       string                                `json:"Status" xml:"Status"`
-	Size         int                                   `json:"Size" xml:"Size"`
+	Size         int64                                 `json:"Size" xml:"Size"`
 	Duration     float64                               `json:"Duration" xml:"Duration"`
 	Description  string                                `json:"Description" xml:"Description"`
 	CreationTime string                                `json:"CreationTime" xml:"CreationTime"`
