+++ services/vod/struct_video_in_get_video_list.go
@@ -21,7 +21,7 @@ type VideoInGetVideoList struct {
 	Title            string                  `json:"Title" xml:"Title"`
 	Tags             string                  `json:"Tags" xml:"Tags"`
 	Status           string                  `json:"Status" xml:"Status"`
-	Size             int                     `json:"Size" xml:"Size"`
+	Size             int64                   `json:"Size" xml:"Size"`
 	Duration         float64                 `json:"Duration" xml:"Duration"`
 	Description      string                  `json:"Description" xml:"Description"`
 	CreateTime       string                  `json:"CreateTime" xml:"CreateTime"`
@@ -29,7 +29,7 @@ type VideoInGetVideoList struct {
 	ModificationTime string                  `json:"ModificationTime" xml:"ModificationTime"`
 	CreationTime     string                  `json:"CreationTime" xml:"CreationTime"`
 	CoverURL         string                  `json:"CoverURL" xml:"CoverURL"`
-	CateId           int                     `json:"CateId" xml:"CateId"`
+	CateId           int64                   `json:"CateId" xml:"CateId"`
 	CateName         string                  `json:"CateName" xml:"CateName"`
 	StorageLocation  string                  `json:"StorageLocation" xml:"StorageLocation"`
 	AppId            string                  `json:"AppId" xml:"AppId"`
